+++ src/Abp.Zero.Common/Authorization/Users/AbpUserPermissionCacheItemInvalidator.cs
@@ -1,6 +1,8 @@
 using Abp.Dependency;
+using Abp.Domain.Repositories;
 using Abp.Events.Bus.Entities;
 using Abp.Events.Bus.Handlers;
+using Abp.Organizations;
 using Abp.Runtime.Caching;
 
 namespace Abp.Authorization.Users
@@ -10,13 +12,16 @@ namespace Abp.Authorization.Users
         IEventHandler<EntityChangedEventData<UserRole>>,
         IEventHandler<EntityChangedEventData<UserOrganizationUnit>>,
         IEventHandler<EntityDeletedEventData<AbpUserBase>>,
+        IEventHandler<EntityChangedEventData<OrganizationUnitRole>>,
         ITransientDependency
     {
         private readonly ICacheManager _cacheManager;
+        private readonly IRepository<UserOrganizationUnit, long> _userOrganizationUnitRepository;
 
-        public AbpUserPermissionCacheItemInvalidator(ICacheManager cacheManager)
+        public AbpUserPermissionCacheItemInvalidator(ICacheManager cacheManager, IRepository<UserOrganizationUnit, long> userOrganizationUnitRepository)
         {
             _cacheManager = cacheManager;
+            _userOrganizationUnitRepository = userOrganizationUnitRepository;
         }
 
         public void HandleEvent(EntityChangedEventData<UserPermissionSetting> eventData)
@@ -41,6 +46,20 @@ namespace Abp.Authorization.Users
         {
             var cacheKey = eventData.Entity.Id + "@" + (eventData.Entity.TenantId ?? 0);
             _cacheManager.GetUserPermissionCache().Remove(cacheKey);
+        }
+
+        public void HandleEvent(EntityChangedEventData<OrganizationUnitRole> eventData)
+        {
+            //get all users in organization unit
+            var users = _userOrganizationUnitRepository.GetAllList(x =>
+                x.OrganizationUnitId == eventData.Entity.OrganizationUnitId && x.TenantId == eventData.Entity.TenantId);
+
+            //delete all users permission cache
+            foreach (var userOrganizationUnit in users)
+            {
+                var cacheKey = userOrganizationUnit.UserId + "@" + (eventData.Entity.TenantId ?? 0);
+                _cacheManager.GetUserPermissionCache().Remove(cacheKey);
+            }
         }
     }
 }
